BusReservationProject

+++ BusReservation/src/main/webapp/index.jsp
@@ -8,6 +8,8 @@
 <head>
 	<meta charset="UTF-8">
 	<title>버스 예약 프로그램</title>
+	
+	<link rel="stylesheet" href="/resources/css/main.css">
 </head>
 <body>
 	
@@ -26,24 +28,35 @@
 						<p>전화번호</p>
 						<input type="text" name="inputPhone">
 					</div>
-					<button>로그인</button>
 					
-					<a href="/signup" class="signup">회원가입</a>
+					<div class="login">
+						<button class="login-btn">로그인</button>
+						<span class="signup">
+							<a href="/signup">회원가입</a>
+						</span>
+					</div>
 				</form>
 			</c:when>
 			
 			<%-- 로그인 했다면 현재 로그인한 사람의 예약 정보 출력 --%>
 			<c:otherwise>
 				<h1>로그인 성공</h1>
-				<h1>${sessionScope.loginMember.name}님 버스 예약 정보</h1>
+				<h2>${sessionScope.loginMember.name}님 버스 예약 정보</h2>
 				
  				<c:choose>
 					<%-- 예약 정보가 없을 때 --%>
 					<c:when test="${empty sessionScope.reserveInfo.departures}">
 						<h2>예약 정보가 존재하지 않습니다.</h2>
-						
-						<a href="/reserveBus">버스 예약하기</a>
-						<a href="/logout">로그아웃</a>
+					
+						<span class="container">
+							<span class="reserve">
+								<a href="/reserveBus">예약하기</a>
+							</span>
+	
+							<span class="logout">
+								<a href="/logout">로그아웃</a>
+							</span>
+						</span>
 											
 
 					</c:when>
@@ -67,22 +80,30 @@
 						</p>
 						
 						<%-- 예약 정보 수정/삭제 --%>
-						<a href="/logout">로그아웃</a>
-						<a href="/updateReserve">예약 변경</a>
-						<a href="/deleteReserve">예약 취소</a>
+						<span>
+							<a href="/logout">로그아웃</a>
+						</span>
+
+						<span>
+							<a href="/reserveBus">예약 변경</a>
+						</span>
+
+						<span>
+							<a href="/deleteReserve" onclick="return confirm('정말 예약취소하시겠습니까?')">예약 취소</a>
+						</span>
 					</c:otherwise>
 				
 				</c:choose>
 				
 			</c:otherwise>
 		</c:choose>
-	</main>
-	
-	<c:if test="${not empty sessionScope.message}">
-		${message}
 		
-		<c:remove var="message" scope="session"/>
-	</c:if>
+		<c:if test="${not empty sessionScope.message}">
+			${message}
+			
+			<c:remove var="message" scope="session"/>
+		</c:if>
+	</main>
 	
 </body>
 </html>
